Cambios ejercicios proyecto

+++ introjava/src/Ejercicios_proyecto/Promedio_de_varios_numeros.java
@@ -6,7 +6,7 @@ public class Promedio_de_varios_numeros {
 
     public static void main(String[] args) {
         Scanner datos = new Scanner(System.in);
-        int num;
+        double num;
         double prom = 0;
         double suma = 0;
         int cont = 0;
@@ -17,7 +17,7 @@ public class Promedio_de_varios_numeros {
             res = res.toLowerCase(); // toLoweCase sirve para cambiar una respuesta de minuscula a mayuscula
             if (res.equals("si")) {
                 System.out.println("Ingrese la Nota:");
-                num = datos.nextInt();
+                num = datos.nextDouble();
                 cont++;
                 suma = (suma + num);
                 prom = suma / cont;
